+++ libs/cmcd/src/CmcdData.ts
@@ -4,6 +4,7 @@ import type { Cmcd } from './Cmcd.ts'
  * All CMCD data types combined.
  *
  * @deprecated Use {@link Cmcd} instead.
+ *
  * @public
  */
 export type CmcdData = Cmcd;