+++ libs/cmcd/CHANGELOG.md
@@ -22,7 +22,7 @@ and this project adheres to
 - Removed deprecated `CmcdData` type alias (use `Cmcd` instead)
 - Removed deprecated `CmcdRequestData` type alias (use `CmcdRequest` instead)
 
-## [2.0.0] - 2026-01-27
+## [2.0.0] - 2026-02-02
 
 ### Added
 
@@ -53,7 +53,7 @@ and this project adheres to
 - Convert to mono-repo ([#238](https://github.com/streaming-video-technology-alliance/common-media-library/issues/238))
 - Produce single bundled export for each package ([#260](https://github.com/streaming-video-technology-alliance/common-media-library/issues/260))
 
-[Unreleased]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cmcd-/cmcd-/cmcd-/cmcd-v2.0.0...HEAD
+[Unreleased]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cmcd-v2.0.0...HEAD
 [2.0.0]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cmcd-v1.0.2...cmcd-v2.0.0
 [1.0.2]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cmcd-v1.0.1...cmcd-v1.0.2
 [1.0.1]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/cmcd-v1.0.0...cmcd-v1.0.1